fix build error when build without `crypto` feature
